requirements.txt
africastalking==1.2.3
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
base58==2.1.0
billiard==3.6.4.0
CacheControl==0.12.6
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels==2.4.0
channels-redis==2.4.2
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==3.4.7
daphne==2.5.0
decorator==4.4.2
Deprecated==1.2.13
Django==3.1.5
django-cors-headers==3.10.0
django-filter==2.4.0
djangorestframework==3.12.2
ecdsa==0.16.1
firebase-admin==4.5.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
httplib2==0.19.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.7.0
incremental==21.3.0
infobip-api-python-client==3.0.1
k2-connect==1.2.2
kombu==5.0.2
Markdown==3.3.4
msgpack==0.6.2
numpy==1.19.5
packaging==20.9
pandas==1.1.5
passlib==1.7.4
#pkg-resources==0.0.0
priority==1.3.0
prompt-toolkit==3.0.18
proto-plus==1.15.0
protobuf==3.15.6
#psycopg2==2.8.6
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2020.5
redis==3.5.3
requests==2.25.1
rncryptor==3.3.0
rsa==4.7.2
schema==0.7.4
sendgrid==6.6.0
service-identity==18.1.0
simplejson==3.17.6
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
twilio==7.3.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
validators==0.18.2
vine==5.0.0
vonage==2.5.5
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.4.1
zope.interface==5.4.0
